+++ extend_ai_toolkit/openai/toolkit.py
@@ -1,22 +1,26 @@
-from typing import List
-
 from agents import FunctionTool
-from pydantic import PrivateAttr
 
-from extend_ai_toolkit.shared import Configuration, ExtendAgentToolkit, Agent
-from .tool import OpenAITool
+from extend_ai_toolkit.shared import (
+    Agent,
+    AgentToolkit,
+    Configuration,
+    ExtendAPI,
+    Tool
+)
+from .extend_tool import ExtendTool
 
 
-class ExtendOpenAIToolkit(ExtendAgentToolkit[FunctionTool, OpenAITool]):
-    _tools: List[FunctionTool] = PrivateAttr(default=[])
+class ExtendOpenAIToolkit(AgentToolkit[FunctionTool]):
 
     def __init__(
             self, api_key: str, api_secret: str, configuration: Configuration = Configuration.allTools()
     ):
         super().__init__(
-            tool_class=OpenAITool,
             agent=Agent.OPENAI,
             api_key=api_key,
             api_secret=api_secret,
             configuration=configuration
         )
+
+    def tool_for_agent(self, api: ExtendAPI, tool: Tool) -> FunctionTool:
+        return ExtendTool(api, tool)
